+++ Editor/Core/Shared/DependencyInversion/AbstractRuntimeServices.Gen.cs
@@ -61,6 +61,6 @@ namespace Unity.Services.Core.Editor.Shared.DependencyInversion
             return scope;
         }
 
-        internal abstract void Register(ServiceCollection collection);
+        public abstract void Register(ServiceCollection collection);
     }
 }