+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.12.0] - 2023-10-20
+### Added
+- Added service registry and new package initialization interface for enabling instances and editor-time services.
+
+### Fixed
+- Resolved issue with `IAccessTokens` reaching rate limit.
+
 ## [1.11.0] - 2023-07-31
 
 ### Added
@@ -14,8 +21,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Implement JsonConverters to prevent NotImplementedException when converters are invoked by reflection
-
-## [1.10.2-pre.24] - 2023-07-20
 
 ## [1.10.2-pre.2] - 2023-06-14